Fixed assignment 3 bug.

+++ src/segment/Record.cpp
@@ -5,7 +5,7 @@ Record::Record(Record&& t) : len(t.len), data(t.data) {
    t.len = 0;
 }
 
-Record::Record(unsigned len, const char* const ptr) : len(len) {
+Record::Record(uint32_t len, const char* const ptr) : len(len) {
    data = static_cast<char*>(malloc(len));
    if (data)
       memcpy(data, ptr, len);
@@ -24,18 +24,18 @@ Record::~Record() {
 }
 
 char* Record::serialize() const {
-	auto bufferSize = sizeof(uint16_t) + getLen();
+	auto bufferSize = sizeof(uint32_t) + getLen();
 	char* buffer = new char[bufferSize];
-	memcpy(buffer, &len, sizeof(uint16_t));
-	memcpy(buffer+sizeof(uint16_t), getData(), getLen());
+	memcpy(buffer, &len, sizeof(uint32_t));
+	memcpy(buffer+sizeof(uint32_t), getData(), getLen());
 	return buffer;
 }
 
 uint16_t Record::serializedSize() const {
-	return sizeof(uint16_t) + getLen();
+	return sizeof(uint32_t) + getLen();
 }
 
 Record* Record::deserialize(char* serialized) {
-	uint16_t len = *((uint16_t*) serialized);
-    return new Record(len, serialized + sizeof(uint16_t));
+	uint32_t len = *((uint32_t*) serialized);
+    return new Record(len, serialized + sizeof(uint32_t));
 }
